requirements.txt
langchain
python-dotenv
streamlit
openai
wikipedia
youtube-transcript-api
pytube